backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
alembic==1.15.2
pydantic[email]==2.11.1
pydantic-settings==2.8.1
python-jose[cryptography]==3.4.0
passlib==1.7.4
python-multipart==0.0.20
redis==5.2.1
celery==5.5.1
cryptography==44.0.2
gitpython==3.1.43
transformers @ git+https://github.com/huggingface/transformers.git@65dc261512cbdb1ee72b88ae5b222f2605aad8e5
torch==2.6.0
librosa==0.10.2.post1
soundfile==0.13.1
python-dotenv==1.1.0
